fix: log level in processJSONHandler
The log level in processJSONHandler was incorrectly set to Debug.
Since this logs a warning about mismatching content types, the log
level should be Warning instead.

This commit changes the log level to Warning when logging the message:

"the "Content-Type" of the response header is not of the "json" type"

The Debug level should only be used for diagnosing issues during
development. Logging warnings about unexpected behavior should use
the Warning log level instead.

This fixes a bug where json parsing warnings were being logged at
the wrong severity level.